very clean cpu fuser
